Factor ValueNodeView out of StaticScope.
Remove unused support in StaticScope for a scope to have multiple parents.
