+++ explorer/ast/expression.h
@@ -17,8 +17,8 @@
 #include "explorer/ast/bindings.h"
 #include "explorer/ast/element.h"
 #include "explorer/ast/paren_contents.h"
-#include "explorer/ast/static_scope.h"
 #include "explorer/ast/value_category.h"
+#include "explorer/ast/value_node.h"
 #include "explorer/common/arena.h"
 #include "explorer/common/source_location.h"
 #include "llvm/ADT/ArrayRef.h"
