+++ explorer/ast/pattern.h
@@ -13,8 +13,8 @@
 #include "explorer/ast/ast_node.h"
 #include "explorer/ast/ast_rtti.h"
 #include "explorer/ast/expression.h"
-#include "explorer/ast/static_scope.h"
 #include "explorer/ast/value_category.h"
+#include "explorer/ast/value_node.h"
 #include "explorer/common/source_location.h"
 #include "llvm/ADT/ArrayRef.h"
 #include "llvm/ADT/STLFunctionalExtras.h"
